Refactoring the menu
