+++ asset-manifest.json
@@ -2,8 +2,8 @@
   "files": {
     "static/js/0.36973d74.chunk.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/0.36973d74.chunk.js",
     "static/js/0.36973d74.chunk.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/0.36973d74.chunk.js.map",
-    "main.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.b8438aad.chunk.js",
-    "main.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.b8438aad.chunk.js.map",
+    "main.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.ead19389.chunk.js",
+    "main.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.ead19389.chunk.js.map",
     "runtime~main.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/runtime~main.ad097248.js",
     "runtime~main.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/runtime~main.ad097248.js.map",
     "index.php": "/react-wp/wp-content/themes/inu-v2-b/index.php"
