+++ react-src/src/components/Category/index.js
@@ -1,6 +1,2 @@
 export { default } from "./Category";
-export {
-  CategoryPropTypes,
-  CategoryDefaultProps,
-  categoryToMenuItem
-} from "./Category";
+export { CategoryPropTypes, CategoryDefaultProps } from "./Category";
