+++ react-src/src/components/MenuItem/MenuItem.stories.js
@@ -4,6 +4,6 @@ import { storiesOf } from "@storybook/react";
 import MenuItem from "./MenuItem";
 import description from "./MenuItem.md";
 
-storiesOf("MenuItem", module).add("Overview", () => <MenuItem />, {
+storiesOf("MenuItem", module).add("With default props", () => <MenuItem />, {
   notes: { markdown: description }
 });
