+++ react-src/src/components/MenuItem/index.js
@@ -1,6 +1,2 @@
 export { default } from "./MenuItem";
-export {
-  MenuItemPropTypes,
-  MenuItemDefaultProps,
-  createMenuItems
-} from "./MenuItem";
+export { MenuItemPropTypes, MenuItemDefaultProps } from "./MenuItem";
